Fixes Finished
Due to Godot 4.2 engine bug, when you change a script name in terms of capitalization only, the tscn files bug out because it does not apply the change. You have to change more than just capitalization, as a result I had to go through this entire project and fix all the .tscn connections.